requirements.txt
docx2pdf
qrcode
pillow